// dashboard/src/client/models/resource-upload-part.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ResourceUploadPart {
    'uploadId': string;
    'partNumber': number;
    'receivedBytes'?: number | null;
    'uploadedBytes': number;
    'nextPart': number;
    'status': ResourceUploadPartStatusEnum;
}

export const ResourceUploadPartStatusEnum = {
    Pending: 'pending',
    Completed: 'completed'
} as const;

export type ResourceUploadPartStatusEnum = typeof ResourceUploadPartStatusEnum[keyof typeof ResourceUploadPartStatusEnum];